+++ md5.h
@@ -20,12 +20,13 @@
 #define MD5_H
 
 // MD5 provides a 16 byte hash
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct
 {
-	uint64_t len;
-	uint8_t  bufLen;
+	size_t   len;
+	size_t   bufLen;
 	uint8_t  buffer[64];
 	uint32_t A;
 	uint32_t B;
@@ -35,9 +36,9 @@ typedef struct
 
 void MD5Init  (MD5_CTX* md5);
 void MD5Block (MD5_CTX* md5, const uint8_t block[64]);
-void MD5Update(MD5_CTX* md5, const uint8_t* data, uint64_t len);
+void MD5Update(MD5_CTX* md5, const uint8_t* data, size_t len);
 void MD5Final (MD5_CTX* md5, uint8_t dst[16]);
 
-void MD5(uint8_t dst[16], const uint8_t* src, uint64_t slen);
+void MD5(uint8_t dst[16], const uint8_t* src, size_t slen);
 
 #endif
