+++ sha512.h
@@ -22,22 +22,23 @@
 // SHA-2: SHA-384/SHA-512
 // SHA512 provides a 64 byte hash
 // SHA384 provides a 48 byte hash
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct
 {
-	uint64_t len; // TODO
-	uint8_t  bufLen;
+	size_t   len;
+	size_t   bufLen;
 	uint8_t  buffer[128];
 	uint64_t H[8];
 } SHA512_CTX;
 
 void SHA512Init  (SHA512_CTX* sha512);
 void SHA512Block (SHA512_CTX* sha512, const uint8_t block[128]);
-void SHA512Update(SHA512_CTX* sha512, const uint8_t* data, uint64_t len);
+void SHA512Update(SHA512_CTX* sha512, const uint8_t* data, size_t len);
 void SHA512Final (SHA512_CTX* sha512, uint8_t dst[64]);
 
-void SHA512(uint8_t dst[64], const uint8_t* src, uint64_t slen);
+void SHA512(uint8_t dst[64], const uint8_t* src, size_t slen);
 
 
 
@@ -45,9 +46,9 @@ typedef SHA512_CTX SHA384_CTX;
 
 void SHA384Init  (SHA384_CTX* sha384);
 void SHA384Block (SHA384_CTX* sha384, const uint8_t block[128]);
-void SHA384Update(SHA384_CTX* sha384, const uint8_t* data, uint64_t len);
+void SHA384Update(SHA384_CTX* sha384, const uint8_t* data, size_t len);
 void SHA384Final (SHA384_CTX* sha384, uint8_t dst[48]);
 
-void SHA384(uint8_t dst[48], const uint8_t* src, uint64_t slen);
+void SHA384(uint8_t dst[48], const uint8_t* src, size_t slen);
 
 #endif
